wifi: ConnectScan: Skip SetMACRandomize when not supported.
This test expects no MAC randomization so that it can filter the
packets with MAC address. On the devices that MAC randomization is not
supported, shill will not enable the option in wpa_supplicant, and
wpa_supplicant by default will trigger scans without setting
NL80211_SCAN_FLAG_RANDOM_ADDR.

With this assumption, we can relax hwdep.WifiMACAddrRandomize and skip
setting MAC randomization on the devices that don't support it.

BUG=b:169117094
TEST=`tast run $DUT wifi.ConnectScan.*` on elm, mickey DUT

Change-Id: Id7dbbe8aea467c8de5feb7f1d5b0a4118158a4a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2434029
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>