wifi: ConnectScan: Skip SetMACRandomize when not supported.
This test expects no MAC randomization so that it can filter the
packets with MAC address. On the devices that MAC randomization is not
supported, shill will not enable the option in wpa_supplicant, and
wpa_supplicant by default will trigger scans without setting
NL80211_SCAN_FLAG_RANDOM_ADDR.

With this assumption, we can relax hwdep.WifiMACAddrRandomize and skip
setting MAC randomization on the devices that don't support it.

BUG=b:169117094
TEST=`tast run $DUT wifi.ConnectScan.*` on elm, mickey DUT

Change-Id: Id7dbbe8aea467c8de5feb7f1d5b0a4118158a4a4