scanner: Add EnumerateIPPUSB test
This test sets up several virtual USB printers and checks that only the
correct subset is included in lorgnette's ListScanners response.

They have to be checked one at a time instead of all at once, because
only one instance of virtual-usb-printer can be running at a time.
However, this is done as a table-driven test instead of a parameterized
test because they should all logically pass or fail together.

BUG=b:159484077
TEST=tast run scanner.EnumerateIPPUSB

Change-Id: Ie668497ab19f24e411df4f94396b1ab2e7809abf