scanner: Refactor lorgnette wrappers into shared package
We want to add more tests that will also need to talk to lorgnette over
dbus.  Refactor out the code that manages the dbus connection and remote
calls so this doesn't have to be duplicated in the new test.  Other than
cleaning up function parameters, the refactored functions behave exactly
the same as the previous non-shared versions.

BUG=b:159484077
TEST=tast run scanner.*

Change-Id: I7e0fd079ee57eca2f7ca4aeab61a5c3fe97234e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2437035
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>