tast-tests: Add wait for iwlwifi driver error dump.
This tests the kernel iwlwifi error collector by inducing an artificial
iwlwifi error and verifying that the expected files are created in the
crash.SystemCrashDir. Some DUTs produce a driver error instead of HW/FW
error which has a different format. I added another wait for the driver
error dump.

BUG=chromium:1015106
TEST=run crash.KernelIwlwifiError

Change-Id: Ief983b569cd385b5d938f451a2662113eb7edbe0