crostini: Report more errors in FsCorruption
Some runs of this test appear to be failing due because vsh fails to
connect to the VM. Currently errors from this are ignored because we
expect an error due to reading a corrupted file. Try to more precisely
filter out the errors we expect, and log them even when we think
they're right so it's possible to check this later.

BUG=chromium:1040272
TEST=Ran test

Change-Id: I22661cf36243d47e1d2d6fad7abd7a1620bd58e1