Remove the waiting of CPU temperature
With the new animated URLs, the CPU temperature could stay higher
than the expected value and could lead to other troubles (including
timeout). Actually this would not be necessary much, because
the test won't fail without waiting, and the result numbers will
be unaffected (mostly because the test repeats the scenario
multiple time).

BUG=none
TEST=tast run ui.OverviewPerf'*' on nocturne

Change-Id: I957c9535f3279435f9bc501440f07d6f61b663c3