tast-tests: Add hwdep.[No]ForceDischarge to PowerAudioPlaybackPerf
On some platforms, force discharging a battery with ectool is not
supported. If so, the battery metric should not be collected.
This CL parameterizes PowerAudioPlaybackPerf to decide whether
battery metrics should be collected on a device or not.
It also prevents calling ectool on devices that do not support
force discharge, which leads the test to fail.

BUG=b:157512297,b:167337854
TEST=tast run <DUT w/o battery> arc.PowerAudioPlaybackPerf* did not
fail

Change-Id: Ifff05d086b854fc2f8e5a53d011ef6415056680a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2442509
Reviewed-by: Matthias Springer <springerm@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Tested-by: Hikaru Nishida <hikalium@chromium.org>
Commit-Queue: Hikaru Nishida <hikalium@chromium.org>