wifi: Fix roaming verification on wifi.RoamDbus.
Changed the roaming verification to wait for BSSID change, and assert
that there was no disconnection. The test should not wait for
intermediate roaming states, because the DUT might not change it's
connected tate.

Added a polling on "VerifyConnection" function after roaming as DHCP
renewal might not yet finish.

BUG=b:159673983
TEST=run tast run wifi.RoamDbus

Change-Id: Ibb7191f04c1ba4602485620fcef6c56eaf651e1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2443586
Tested-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Hsin-chen Chuang <chharry@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>