arc.VideoDecodeAccel: Add option to use software video decoder.
This CL adds the option to use a software video decoder when running
the ARC++/ARCVM video decoder tests. Currently SW decoding is not
enabled by default in any test, but it can be enabled by adding
"DecoderType: video.SoftwareDecoder" to the DecoderTestOptions.

SW decoding can be enabled in the following tests:
- arc.VideoDecodeAccel.*
- arc.VideoDecodeAccelPerf.*

BUG=b:153402533
TEST=tast run DUT arc.VideoDecodeAccel.h264

Change-Id: I77c953a637d2f14d35ec7e3567e7da95838bf23b