wifi.BgscanBackoff: Turn off power save in this test.
When power save is turned on, the ping RTT will depend on the timing of
beacons. Also, the beacon interval (~102 ms) is quite large compared
with our thresholds. We should turn the power save off to obtain more
accurate RTT so that the test can work properly.

BUG=b:168681584
TEST=tast run $DUT wifi.BgscanBackoff*

Change-Id: Ia064fa90c2cc374653270f64dec4abd77dc532d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2448722
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>