ui: New node check approach for policy tests
In policy tests, a node might exist or not depending on the policy value.
The existing functions can only return fast either for existing or
not existing nodes, for the other case they have to timeout.
In the timeout case for these, the error is random, it can be either
a context timeout or ErrNodeDoesNotExist.

BUG=chromium:1133711
TEST=tast run <ip> policy.ArcEnabled
TEST=tast run <ip> policy.ShowHomeButton
TEST=tast run <ip> policy.ShowLogoutButtonInTray
TEST=tast run <ip> policy.TouchVirtualKeyboardEnabled

Change-Id: Icf41a0a3a6d9d2f38a6e10781a944e5944b2b685
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2448495
Tested-by: Gabor Zsolt Magda <gabormagda@google.com>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Commit-Queue: Gabor Zsolt Magda <gabormagda@google.com>