tast-tests: Add test for force installed policy
Add a test for ExtensionInstallForceList user
policy. The test verifies that extensions specified in the policy
are installed and cannot be removed when "Remove from Chrome"
button is clicked on the extensions page. Used the Platform Keys
Extension which is also used in the corresponding autotest.

BUG=chromium:1125563
TEST=tast run <ip> policy.ExtensionInstallForceList

Change-Id: I7f94828e65eb3a6bcf328314c3b21a1f4a5bc8cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2449553
Tested-by: Swapnil Gupta <swapnilgupta@google.com>
Commit-Queue: Swapnil Gupta <swapnilgupta@google.com>
Reviewed-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>