Stop running crash.ECCrash on some broken boards.
The EC collector doesn't work on several boards due to the boards not
having EC firmware, the EC firmware not supporting the `crash` command,
or a to-be-investigated issue on Nocturne (crbug.com/1135798).

Disable the test on broken boards.

Names of boards based on device-config.txt in stainless files page.

BUG=chromium:1135798,chromium:1123716,chromium:1123191
TEST=`tast run -var servo=chromeos6-row10-rack6-labstation1.cros:9997 \
     chromeos6-row10-rack6-host5.cros.corp.google.com crash.ECCrash`;
     verified test skip.

Change-Id: If35770617f69093885bd557f8b945871a00e4b93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2453854
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Alexander Hartl <alexanderhartl@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>