Reform the run with traces to ensure cleanup.
The reported test log indicates that sometimes the test can fail
during "StartTracing" request while the tracing request has actually
arrived to the browser and so the tracing session started.

That's not great. This CL makes sure that a tracing session is
stopped;
- if there is not enough time in the context, it just skips
  the tracing run.
- even when StartTracing returns an error, the tracing session
  may have started.  So StopTracing should be invoked.
- increase the duration for StopTracing to 2 seconds just in case

I also think that, when tracing is failed to record, it wouldn't
be a big problem as it can report the same numbers. Therefore
some of s.Fatal / Error is replaced by s.Log, that way it can
say why tracing is not recorded, but doesn't disturb the result.

BUG=chromium:1135836
TEST=made sure normal test runs with ui.OverviewScrollPerf

Change-Id: Iab4674e20f660c18dbdea99280a09aee6fccf6d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2453695
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>