crostini: Increase timeout for FsCorruption to 15 minutes
We are seeing cases where this test times out because of how long it
takes to copy the whole disk image around. While we try to make a
sparse copy, which is faster, on some devices even this seems to be
too slow.

We can also reduce the number of disk copies in the test run from 3 to
2, since after the last call to testOverwriteAtOffsets we can rename
the backup of the VM disk over the image rather then copying the data,
which is much faster.

BUG=chromium:1040272
TEST=Ran test

Change-Id: Idbc49f766d766766291ff19e1eb59524ff021011