crostini: Reduce copies in FsCorruption
We are seeing cases where this test times out because of how long it
takes to copy the whole disk image around. While we try to make a
sparse copy, which is faster, on some devices even this seems to be
too slow.

However, we can reduce the number of disk copies in the test run from
3 to 2. After the last call to testOverwriteAtOffsets we can rename
the backup of the VM disk over the image rather then copying the data,
which is much faster.

BUG=chromium:1040272
TEST=Ran test

Change-Id: Idbc49f766d766766291ff19e1eb59524ff021011
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2454961
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>