arc.VideoEncodeAccel: Add option to use software video encoder.
This CL adds the option to use a software video encoder when running
the ARC++/ARCVM video encoder tests. Currently SW encoding is not
enabled by default in any test, but it can be enabled by adding
"EncoderType: video.SoftwareEncoder" to the ArcTestOptions.

SW encoding can be enabled in the following tests:
- arc.VideoEncodeAccel.*
- arc.VideoEncodeAccelPerf.*

BUG=b:169626661
TEST=tast run DUT arc.VideoEncodeAccel.h264_192p_i420 on grunt

Change-Id: I2a256cc53ccbad2849a15069a38b9a7453d7c10a