tast-tests: stabilize VirtualKeyboardFloat
https://tastboard.corp.google.com/test/inputs.VirtualKeyboardFloat.unstable

VirtualKeyboardFloat test failed intermittently and throws out error as below:
fail to wait for virtual keyboard shown: a javascript remote object was not return

The main reason is that VK uses different extension instance in login profile and user profile. So it needs to wait for login extension to disappear before performing any test.

To solve this problem:
1. Adding vkb background check right after user login in chrome.go.
2. Create vkEnabled option for chrome instance.
3. Refactor existing tests using chrome option instead of extra args.

BUG=b:170260568
TEST=inputs.VirtualKeyboardFloat*

Change-Id: I6233ccb2ef0a5f5437c6fa5c424605d5d20d17a7