tast : Enable verbose logging upon GAIALogin
When running Chrome with the gaiaLogin or fakeLogin Option it will
additionally enable verbose logging on some enrollment related files.
Chrome OS OAC suspects that some forced enrollment issues might
be causing the login to get stuck. Once the cause of the issue
is found, the verbose logging will be disabled again.

BUG=chromium:1109176
TEST=tast run DUT assistant.StartupInSignedInMode (uses GAIALogin)

Change-Id: If869b7e10712dbf84c841ff1e24af37f97eedc36