crostini.FsCorruption: Don't ignore errors in deferred statements
We do several important and potentially slow cleanup steps in deferred
statements, which by default means any errors will be
discarded. Ignoring these errors is problematic in general, but
especially when the error is a context expiry which will cause the
point at which the test timed out to be misreported.

We can preserve these errors by deferring an anonymous function that
calls the real method and either stores error results in a named
return value, or logging them.

BUG=chromium:1040272
TEST=Ran test

Change-Id: I61a0082e19f05feca3edda2b8091fa90d554cbc3