crostini: fix context expiry preventing error messages
Many error messages currently look like the following:
[Precondition failure] Failed to install Crostini: failed to install Crostini from UI: Install(): Step 3: uig.WaitUntilDescendantGone({Role:rootWebArea Name: ClassName: Attributes:map[name:^Set up Linux \(Beta\) on your ] State:map[]}, 10m0s): uig.Root().WaitUntilDescendantGone({Role:rootWebArea Name: ClassName: Attributes:map[name:^Set up Linux \(Beta\) on your ] State:map[]}, 10m0s): context deadline exceeded; last error follows: node still exists, [Precondition failure] Panic: runtime error: invalid memory address or nil pointer dereference

The code that should provide a better error message is not being run as
the context has already expired.

Instead, shorten the context for the main operation to allow time for
error checking if the context deadline is exceeded.

Also fix the nil pointer dereference when the context does expire.

BUG=None
TEST=Ran crostini.Basic.artifact on hatch

Change-Id: Id671d020d9e5b52c804a57dd8b98d0f6b8e2ab0a