tast-tests/webrtc: Stagger parallel encoder startup
Add a small delay to stagger startup of parallel encoders. The renderer
process enumerates encoding capabilities asynchonously on startup. This
leads to the case where the renderer doesn't yet know about HW encoding
when JS tries to create an encoder. A small delay already exists in the
test between creating a renderer and starting the encoder, but this is
less effective when the system load increases.

Staggering encoders gives each renderer more time to finish loading
before starting an encode.

BUG=b:168834129
TEST=tast run <DUT> 'webrtc.MediaRecorderMulti.*'

Change-Id: I7891a29b00e5a33bca5115e700eef5f22263cf58