graphics: Set glxgears to specific size.
graphics.TraceReplay.glxgears* is slightly different from its private traces
brother as they exercise the resize code path. It leads to some problems
which are not caught by our glxgears tests.
Modify the precondition and glxgears to have 4 GB of disk size
so that we always exercise the same code path.

BUG=b:170224038
TEST=tast run $DUT graphics.TraceReplay.glxgears*

Cq-Depend: chrome-internal:3324327
Change-Id: I9ad0ced9a4b874d67ec8ebe7d489420859e7046b