tast-tests: stabilize LaunchReleaseNotesFromSettings
2 errors are thrown out quite often as below:
1. Failed to click See whats new: TypeError: Cannot read property 'left' of undefined
2. Failed to verify "what's new" tab rendering: failed to find help app: cdp.Runtime: CallFunctionOn: context deadline exceeded

The main reason is that when clicking "See whats new" menu, the ui node is not yet stable. It might be refreshed in a11y tree for example.
Solution: Add FindAndClickStableNode function in automation.go to wait for node stable before clicking it.

Why it is not happening on LaunchHelpAppFromSettings, my guess is
1. In a very short time period, some other page / sub elements happen to have the same node params `See whats new` (it might not be visible)
2. By using tast precondition, LaunchReleaseNotesFromSettings is executed after LaunchHelpAppFromSettings. In theroy,  between test cases all opened targets should be closed, but it can not be guaranteed old a11y is completely cleaned up.

Either of them should be fixed by using StableFindAndClick. So I would not go for deeper investigation unless it happens again.

Tested locally 100 / 100 passed.

BUG=b:170158471
TEST=apps.*

Change-Id: I3c7d545174744b39758f896d5e3e2e6e11b8daa0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2451831
Tested-by: Jun Sheng <shengjun@chromium.org>
Reviewed-by: Rachel Carpenter <carpenterr@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Commit-Queue: Jun Sheng <shengjun@chromium.org>