wifi: Use wpa_supplicant 'Reattach' for WiFiService 'Reassociate'
The Autotest code that this is a port of actually used the 'Reattach'
method, so this change makes `wifi.Reassociate` a more accurate port
of `network_WiFi_Reassociate`.

The `Reattach` method "attempt[s] reassociation to the current BSS",
whereas `Reassociate` omits the "current BSS" bit. In most cases the
distinction doesn't really matter and the DUT just takes a couple of
extra seconds doing a scan first, but on Dalboz (and I suspect any
RTL8822CE device) the scan takes long enough that the test fails.

As mentioned in the bug, the fact that Dalboz takes >10s to do a
scan when using the `Reassociate` method is concerning, but also
"Reattach is the command we actually use in Shill under certain
scenarios", so that's the one we ought to be testing here.

BUG=b:169217691
TEST=tast run wifi.Reassociate on Helios and Dalboz DUTs

Change-Id: I9c4f557474309121134fc33598a1b14df885fca2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2461672
Tested-by: Will Donnelly <wgd@google.com>
Commit-Queue: Will Donnelly <wgd@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>