tast-tests: simplify platform.CrosHealthdProbeCPUInfo validation
Previously the exact number of lines for each element of the CpuInfo was
validated for every block. This change modifies the output line number
validation to only account for the block it is validating. This fixes an
issue where CPUs without CStates were eronously failing the test.

BUG=chromium:1134670
TEST=tast run veryon_tiger-dut platform.CrosHealthdProbeCPUInfo

Change-Id: I020476619e30bb6d2be6ef09042d289f023be195
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2461673
Tested-by: Trent Begin <tbegin@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Trent Begin <tbegin@chromium.org>