arc.MouseInput: Expect hover after up event
The behavior in vanilla Android for mice is to send a ACTION_HOVER_MOVE
event immediately following an ACTION_UP to let apps know that the mouse
is now hovering.

We update the test to reflect that change.

This also removes the need to sleep in the test.

BUG=None
TEST=tast run DUT arc.MouseInput

Change-Id: Ib8fdd56a23b62aa7b0e0dacef8b1dd98e6693d2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2462706
Reviewed-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Prabir Pradhan <prabirmsp@chromium.org>
Auto-Submit: Prabir Pradhan <prabirmsp@chromium.org>
Tested-by: Prabir Pradhan <prabirmsp@chromium.org>