Add waiting in WindowResizePerf
ui.WindowResizePerf has been failing on some devices, and looks
like it is due to cpu.WaitUntilIdle -> power.WaitUntilCPUCoolDown
replacement -- I also confirmed that locally reverting back to
cpu.WaitUntilIdle fixes the problem.

By looking into the code and behaviors slightly more, I've come
to realize that it is a lack of waitings for the window animations
caused by setting the window state.  If power.WaitUntilCPUCoolDown
waits too short, the window is still animating and ends up with
dragging at wrong locations. cpu.WaitUntilIdle made a lot more
of waiting in average, so this problem wasn't observed previously.

So, this CL introduces ash.WaitWindowFinishAnimating which should
make a proper waiting.

BUG=chromium:1135365
TEST=ui.WindowResizePerf on auron_paine

Change-Id: Idc1e77d3d1c538f22321fe3af42d33c24862d998