tast-tests: arc.InputCompat: Skip scroll in tablet mode
The test was constantly failing in the lab for kohaku because the DUTs
were in tablet mode when the test was run, and all injected events to
the trackpad were ignored. For purely tablet devices, it was passing
because it created a virtual trackpad device for those devices, but
let's stop that altogether as it's not really testing anything.

BUG=None
TEST=tast run kohaku-DUT arc.InputCompat w/ tablet mode enabled

Change-Id: I453fd144a1487d17bdd053ef7b6eb9785406680a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2462844
Tested-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Commit-Queue: Tetsui Ohkubo <tetsui@chromium.org>