wifi: ConnectScan: Don't fail when no probe request found.
In network_WiFi_VisibleScan, it does not fail when no probe request is
found in pcap. This makes sense as it is possible that DUT already
find the service by receiving the beacon frame from AP. Let's relax the
condition and sync the logic in Autotest.

For network_WiFi_HiddenScan case, we assert both wildcard and service
SSID to be found so if no probe request found, the test will still fail.
So the change will not affect its coverage.

BUG=b:169117094
TEST=tast run $DUT wifi.ConnectScan*

Change-Id: Id9c13dcb43bcc959ec4d6c368373e1459369420e