wifi: ConnectScan: Don't fail on malformed probe requests.
In both network_WiFi_VisibleScan and network_WiFi_HiddenScan, they do
not fail when the probe requests in pcap are malformed. This difference
is causing some flakes in the Tast version recently. Let's sync the
behavior in Autotest to unblock the Tast migration task for now.

Also, in order not to lose coverage, a similar test will be added in
upcoming commit so that we can trace if this is a real problem with it.

BUG=b:169117094
TEST=tast run $DUT wifi.ConnectScan*

Change-Id: Ic4a0ee2f1ddbea5c653ca94d762ea2577d5c9d11
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2465611
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>