wifi: Add ProbeReqFormat test.
In previous commit, we're relaxing the condition in wifi.ConnectScan.
However, we find some potential problem in b/169117094#c3. Add this
test for the removed coverage on ensuring that the probe requests
captured is not malformed. If the test fails, it is either a problem
with DUT or the test pcap and both needs to be investigated.

Also as this test is triggering more scans, it is more likely to catch
the problem.

BUG=b:169117094
TEST=tast run $DUT wifi.ProbeReqFormat

Change-Id: I67c7a4f00f78c94ed9c616a216e554ed02853c2d