wifi: Fix a conflict in TFRouter change.
The PersistenceBluetoothSansWifi test fails with "failed to connect to
the router : dial tcp: missing address" recently.

In crrev.com/c/2412333, the TFRouter now can take multiple routers and
empty string is no longer treated the same as no TFRouter specified.
Update the usage to fix the test failure.

BUG=b:161902990
TEST=`tast run $DUT wifi.PersistenceBluetoothSansWifi` and the test
does not fail with "failed to connect to the router".

Change-Id: Ifa2e2804436524dcff0cd3629691e24e2612da15