wificell: hostapd: Assign random BSSID when not specified.
In b/161648841, we found that there's an issue when BSSID is re-used to
configure a difference network, driver might fail to distinguish them
in scan results if the old one is not expired.

This is causing some test failures and blocks the Tast migration task.
Add a work-around to use the unique MAC address logic as default value
for BSSID. For the BSSID re-used case, we can write new tests for the
specific cases to maintain the test coverage.

BUG=b:161648841
TEST=tast run $DUT wifi.*
TEST=Run the test in crrev.com/c/2331112.

Change-Id: I97dddf90c7f9b747adc7c0799ad01647db19449c