wificell: hostapd: Assign random BSSID when not specified.
In b/161648841, we found that there's an issue when BSSID is re-used to
configure a difference network, driver might fail to distinguish them.

This is causing some test failures and blocks the Tast migration task.
Add a temporary work-around to re-use the unique MAC address logic as
default value for BSSID. We should revert this change after the problem
is fixed.

BUG=b:161648841
TEST=tast run $DUT wifi.*

Change-Id: I97dddf90c7f9b747adc7c0799ad01647db19449c