wifi: Add CSALeaveChannel test to wificell_cq.
The Autotest counterpart of this test is in wificell_cq so before we
remove the Autotest version, we would want the Tast version to be
included in wificell_cq so that we don't lose coverage.

The test is already in wifi_matfunc and according to [1], it has >99%
pass rate so it should be good to be in wificell_cq.

[1] http://go/cros-green-stainless-dashboard?f=test_name:in:tast.wifi.CSALeaveChannel

BUG=b:154879577
TEST=`tast list $DUT '("group:wificell" && wificell_cq &&
!wificell_unstable)'` does include wifi.CSALeaveChannel

Change-Id: I4c13921b80ca616f6fc9207fbdd637ac99e25619