usbprinter: Wait for udev event on shutdown
When a test is done with the virtual USB printer, it kills
virtual-usb-printer and waits for the process to exit, but it doesn't
wait until the kernel finishes cleaning up the USB devices.  If a
subsequent test restarts virtual-usb-printer quickly enough, usbip can
fail to attach to its expected device.

Eliminate this race by waiting for a remove udev event after killing
virtual-usb-printer.  This logic is refactored into usbprinter so that
it doesn't have to be repeated in every test.

BUG=b:159484077
TEST=1300 iterations of scanner.EnumerateIPPUSB on betty

Change-Id: I77f65898736448370075b42402716a04d0b3138f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2471023
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>