tast-tests: Add support to retrieve perf sched output
Scheduling latency of cras will directly effect cras performance. We
found issues with audio glitches recently. This patch adds support to
perf package to measure maximum scheduling latency of a process. The
name of process to monitor has to be passed.

Also make use of it in cras to add a new metric (cras_max_latency_ms).

Currently we don't pass/fail test and the metric is added just to
monitor in crosbolt. In the future, we could make test fail if latency
is, say, > 10ms.

While we're at it, update one of the TODO links to better reflect the
current state -- this test won't crash an aarch64 system, but it still
doesn't pass reliably. If we can't fix the bugs reasonably soon, we
should probably update this to a SoftwareDep, instead of doing
skip-testing here.

BUG=b:166194367, b:194820340
TEST=`tast run ${HOST} audio.CrasPerf.playback`

Change-Id: I1abadc2caadf6cbac9b340c17a0708099fe92fa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2478482
Commit-Queue: Joel Fernandes <joelaf@google.com>
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Reviewed-by: Joel Fernandes <joelaf@google.com>
Tested-by: Joel Fernandes <joelaf@google.com>