tast-tests: Add support to retrieve perf sched output
Scheduling latency of cras will directly effect cras performance. We
found issues with audio glitches recently. This patch adds support to
perf package to measure maximum scheduling latency of a process. The
name of process to monitor has to be passed.

Also make use of it in cras to add a new metric (cras_max_latency_ms).

Currently we don't pass/fail test and the metric is added just to
monitor in crosbolt. In the future, we could make test fail if latency
is say > 10ms.

BUG=b:166194367
TEST=tast -verbose run -build=true joelhome:8023 audio.CrasPerf.playback

Change-Id: I1abadc2caadf6cbac9b340c17a0708099fe92fa3