video.Contents: Improve video bounds detection.
This CL changes how video.Contents finds where a full-screen video is on
a screenshot. Before, we assumed that there were no black strips on the
left or on the right of the video and that the black strip on the top
was of the same size as the strip on the bottom.

On grunt, I noticed that the only black strip was on the right of the
video and it was only 1px wide. This is probably because the video
couldn't be centered on the screen perfectly symmetrically.

Now, we do a more general computation: we independently process each
side. We just start from the edge of the screenshot towards the video
until we find a non-black pixel.

BUG=b:159158467
TEST=Verified bounds on eve and grunt with video.Contents.h264_360p_hw

Change-Id: I120ee8d9eb03cc71b2d7a249fc18121277e09c31