network: Add new subsystem ID for HrP2.
In b/159410254 and crrev.com/c/2260678, 0x6074 is added as another
subsystem ID for HrP2 RSE on nightfury EVT3. However, the known device
list in wlan.DeviceInfo is not updated and the tests using this function
will fail on nightfury EVT3 with unknown device error.

Add the ID to the list so that network.WifiCaps and network.WLANDriver
can pass.

BUG=b:171159634
TEST=`tast run $DUT network.WLANDriver network.WifiCaps` with
chromeos6-row12-rack6-host10 in lab.

Change-Id: Ie0e63d9c3ca1dcd0864b1961dbf8e48f50acce15