memoryuser metrics tasks
Introduces new tasks that can be used to report metrics about which
MemoryTasks have been killed, or failed to start properly.

Moves perf.Metrics into testEnv so that tasks can report metrics if they
wish.

I noticed that MemoryTask.Close is not called unless we run parallel,
but adding CLose caused metrics to break because they were collected
after the close (so all tabs counted as killed). So I thought this would
be a more flexible way to add metrics to the test than the
extraPerfMetrics flag.

BUG=b:154785124
TEST=tast run <DUT> arc.LifecycleChromeOSPerf.vm

Change-Id: I5dca54441c237a09f4db6b0a875498e8c43510ee