tast-tests: Revert Install New Service to Add New Service
The change crrev/c/2422793 was added due to an accessibility
update on the Chrome side causing propagations of nodes to
not update the accessibility tree. This has been fixed in
crrev/c/2446089 and thus reverting the change.

Remove finding the SMB Share window as it's sufficient to
wait for the File share URL to appear.

BUG=chromium:1140884
TEST=tast run <DUT-IP> filemanager.SMB

Change-Id: I7c53121d61aed36ecb9f8c500e69446ffb034a4c