wifi: Add ChannelScanDwellTime test
This is a port of network_WiFi_ChannelScanDwellTime from Autotest.

This test uses frame_sender to transmit a bunch of beacon frames
approximately 1ms apart with sequential SSIDs, triggers a scan,
and then locates the scanned SSIDs in a packet capture to compute
the time interval between the first and last beacons found.

BUG=b:171600107
TEST=tast run $DUT wifi.ChannelScanDwellTime

Change-Id: I3fda02b5116b91230434ae9702f085f077b50909