wifi: OptionalDHCPProperties: Relax the filter for better robustness.
In b/162395242#comment7, we find a case that pcap failed to collect the
DHCP request sent by DUT. However, pcap did collect the one relayed by
AP to broadcast address. In a normal run of this test, there might be
only one DHCP request sent and there's some probability that pcap can
miss it and cause the test flaky. Let's also include the packet that
is relayed by AP (with source address == MAC of DUT) so we can have one
more chance and reduce the possibility of "No DHCP found" failure.

BUG=b:162395242
TEST=tast run $DUT wifi.OptionalDHCPProperties

Change-Id: Ia8fa32911e8d20e4256b0b81d28d674f19bee4f6