network: Add ping option to save output.
In some tests, the detailed log of ping is useful for debug, e.g.
we can know the sequence number of ICMP packets with long latency in
BgscanBackoff test to reference the pcap file. The detailed result is
usually not that useful in test logic, so let's add an option to save
the output for debugging.

Also apply the option to some tests that have long pinging.

BUG=b:168681584
TEST=`tast run $DUT wifi.BgscanBackoff` and check the ping output is
saved in test results.

Change-Id: Ice2ba893126d71677ff3b5abd0c41e7d8068e855
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2491579
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Hsin-chen Chuang <chharry@google.com>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>