wificell: Refactor Router and APIface.
Split the logics used to be in Router.StartAPIface + APIface.start to
Router.StartHostapd + Router.StartDHCP, and move the methods using
ip_runner or iw_runner from APIface to Router.
It is helpful to the tests who need to start the services with more
precise configurations.

Move the busySubnet map to ap_iface.go as a global variable because:
1) It should not be a field in a Router object since there may be more
   than one router in the future.
2) Other assumptions about IP are in ap_iface.go (broadcastIP, ServerIP,
   IP prefix, etc).

BUG=none
TEST=tast run ${DUT} wifi.ConnectionIdentifier wifi.LinkMonitorFailure

Change-Id: I9a47e13f53ab798b123e09fc76384c099ee6eb64