Add printer.TestPPDs test
This test runs cupstestppd on the PPD files in ppds_all.tar.xz.

It also verifies that the foomatic command parses properly.

It runs orders of magnitude faster than the platform_printerPpds
autotests. The total runtime of the test is ~1.5 minutes.

PPD files that fail the test have been removed from the
archive.

BUG=none
TEST=tast run -buildbundle crosint localhost:9222 printer.TestPPDs

Cq-Depend: chromium:2519303
Change-Id: I1b1ff1b49a33dc878ed2c082a4a2d2205b86c5ed