Add printer.TestPPDs test
This test runs cupstestppd as well as the foomatic-rip filter on the
5837 PPD files from the v3 index that pass the tests.

It verifies that the both the PPD file and foomatic shell command in
the PPD file (if present) parse properly.

It runs orders of magnitude faster than the platform_printerPpds
autotests. The total runtime of the test is around 13s (triple when
cacheFoo is disabled).

BUG=none
TEST=tast run -buildbundle crosint localhost:9222 printer.TestPPDs

Cq-Depend: chromium:2519303
Change-Id: I1b1ff1b49a33dc878ed2c082a4a2d2205b86c5ed