tast-tests: Use shorter context and COF for arc.AppLoadingPerf
There is a known libhoudini issue http://b/169367367 that can hang the
test forever and it is not efficient to wait that long.  Based on data
from stainless, all subflow tests should complete within 5 minutes. If
subflow test fails, it will continue to log data for other subflows and
not exit immediately.  Also update APK for UI bug http://b/166352175.

BUG=b:169623350
BUG=b:166352175
BUG=b:153866893
TEST=tast run arc.AppLoading.*

Change-Id: I6998e3c7b1a091a7ee7cd8e6275b6737eb8bdd38