video.Contents: Add tolerance for black strips.
This CL changes video.Contents to incorporate a tolerance when detecting
the black strips that surround a video when in full screen. Before this
CL, the color was expected to be exactly black (0, 0, 0). With this CL,
we allow for a tolerance of 25 (i.e., every component must be less than
25 away from the corresponding black component). This was obtained
empirically: on kukui, the first row of 'black' below the video was
observed to be actually (20, 1, 22), presumably due to blending
artifacts that don't have a real visual impact. Without this tolerance,
the bounds of the video can't be correctly detected.

Note that the minimum distance between pure black and any of the colors
in the test video should be 128, so a tolerance of 25 should be okay
(there might be blending artifacts that reduce this minimum distance,
but 25 is still fairly low).

BUG=b:159158467
TEST=Verified detected bounds on a kukui for h264_360p_composited_hw

Change-Id: I30296157d38c46ca0df6a70719ae1390370465ba